Don't assume shared library is installed relative to files in lib
This currently works, but it's planned to eventually stop doing it.
Furthermore, there's  a `.gemrc` configuration
(install_extension_in_lib) that some OS packagers set to false to enable
the future behavior, and gems making this assumption don't work with
that.

Honestly, it feels like a very hard change for RubyGems to make smoothly
and that makes me doubt it's worth doing, but this would be the "blessed
way" that should not break regardless of what we end up doing.